+++ feed/urls.py
@@ -1,1 +1,8 @@
 from django.conf.urls.defaults import *
+from views import RecentRecipesFeed
+
+urlpatterns = patterns('',
+
+    (r'^recent/$', RecentRecipesFeed()),
+
+)